+++ src/js/index.js
@@ -1,4 +1,0 @@
-require("../sass/styles.scss");
-const renderSong = require("./lib/render-song.js");
-
-renderSong();