+++ src/typescript/index.ts
@@ -1,0 +1,6 @@
+import "../sass/styles.scss";
+import { renderSong } from "./lib/renderSong";
+import { renderRecipeCalc } from "./lib/recipeCalc";
+
+renderSong();
+renderRecipeCalc();