+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1]
+## [Unreleased]
+### :rocket: Added
+### :pencil: Changed
 
+## [1.0.0]
+### :rocket: Added
+- Recipe Calculator
+### :pencil: Changed
+- Song Infobox styles removed.
+
+## [0.0.1]
 ### :rocket: Added
 
 - Song color coding inside any `.song` html element
 - Song Infobox styles for MediaWiki Template
 
-[Unreleased]: https://github.com/erremauro/CignoApp/compare/v0.0.1...HEAD
+[Unreleased]: https://github.com/erremauro/CignoApp/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/erremauro/CignoApp/releases/tag/v1.0.0
 [0.0.1]: https://github.com/erremauro/CignoApp/releases/tag/v0.0.1